feat: Add utils function to transform users with clerk; Set User function at chat Context update request;
